# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Household Finance API
  version: 0.1.0
  description: |
    HTTP API for the self-hosted household finance app. All authenticated routes expect
    `Authorization: Bearer <JWT>` unless noted. See `docs/API_INDEX.md` for prose topic guides.
servers:
  - url: /
    description: Relative to configured API base URL
...
paths:
...
  /imports/accounts:
    get:
      tags: [imports]
      summary: List financial accounts
      security: [{ bearerAuth: [] }]
    post:
      tags: [imports]
      summary: Create connected account
      security: [{ bearerAuth: [] }]
  /imports/accounts/{accountId}:
    patch:
      tags: [imports]
      summary: Update connected account
      security: [{ bearerAuth: [] }]
  /imports/parser-profiles:
    get:
      tags: [imports]
      summary: List known parser profile ids
      security: [{ bearerAuth: [] }]
  /imports/institutions:
    get:
      tags: [imports]
      summary: Curated institution labels and household custom institutions
      security: [{ bearerAuth: [] }]
  /imports/institutions/custom:
    post:
      tags: [imports]
      summary: Add household custom institution name
      security: [{ bearerAuth: [] }]
  /imports/sessions:
    post:
      tags: [imports]
      summary: Create import session
      security: [{ bearerAuth: [] }]
  /imports/sessions/{sessionId}:
    get:
      tags: [imports]
      summary: Session detail
      security: [{ bearerAuth: [] }]
  /imports/sessions/{sessionId}/status:
    patch:
      tags: [imports]
      summary: Transition session status
      security: [{ bearerAuth: [] }]
  /imports/sessions/{sessionId}/summary:
    get:
      tags: [imports]
      summary: Per-file import summary
      security: [{ bearerAuth: [] }]
  /imports/sessions/{sessionId}/files:
    post:
      tags: [imports]
      summary: Upload files to session (multipart field `files`)
      security: [{ bearerAuth: [] }]
  /imports/sessions/{sessionId}/files/{fileId}:
    patch:
      tags: [imports]
      summary: Bind file to account and parser
      security: [{ bearerAuth: [] }]
  /imports/sessions/{sessionId}/parse:
    post:
      tags: [imports]
      summary: Parse staged files
      security: [{ bearerAuth: [] }]
  /imports/sessions/{sessionId}/canonicalize:
    post:
      tags: [imports]
      summary: Canonicalize into ledger
      security: [{ bearerAuth: [] }]
  /imports/sessions/{sessionId}/undo-import:
    post:
      tags: [imports]
      summary: Roll back session ledger effects
      security: [{ bearerAuth: [] }]
...
  /payslips:
    get:
      tags: [payslips]
      summary: List payslips
      security: [{ bearerAuth: [] }]
  /payslips/upload:
    post:
      tags: [payslips]
      summary: Upload payslip PDF
      security: [{ bearerAuth: [] }]
  /payslips/sniff:
    post:
      tags: [payslips]
      summary: Sniff PDF for parser hints
      security: [{ bearerAuth: [] }]
  /payslips/{id}:
    get:
      tags: [payslips]
      summary: Payslip detail
      security: [{ bearerAuth: [] }]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT